# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Timetracker API Server"
        }
    ],
    "paths": {
...
        "/statuses": {
            "get": {
                "tags": [
                    "Assignment statuses"
                ],
                "summary": "Get list of statuses",
                "description": "Returns list of statuses",
                "operationId": "getStatusesList",
                "parameters": [
                    {
                        "name": "Page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Per page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "completed",
                                                    "created_at": "2021-04-22T16:12:24.000000Z",
                                                    "updated_at": "2021-04-22T16:12:24.000000Z",
                                                    "deleted_at": null
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "in progress",
                                                    "created_at": "2021-04-22T16:12:24.000000Z",
                                                    "updated_at": "2021-04-22T16:12:24.000000Z",
                                                    "deleted_at": null
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assignment statuses"
                ],
                "summary": "Store new status",
                "description": "Returns status data",
                "operationId": "storeStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 3,
                                        "name": "waiting",
                                        "created_at": "2021-04-22T16:12:24.000000Z",
                                        "updated_at": "2021-04-22T16:12:24.000000Z",
                                        "deleted_at": null
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/statuses/{id}": {
            "get": {
                "tags": [
                    "Assignment statuses"
                ],
                "summary": "Get status information",
                "description": "Returns status data",
                "operationId": "getStatusById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 1,
                                        "name": "completed",
                                        "created_at": "2021-04-22T16:12:24.000000Z",
                                        "updated_at": "2021-04-22T16:12:24.000000Z",
                                        "deleted_at": null
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Assignment statuses"
                ],
                "summary": "Update existing status",
                "description": "Returns updated status data",
                "operationId": "updateStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": true
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assignment statuses"
                ],
                "summary": "Delete existing status",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}